Refactor new Helm code into experimental package for now
