+++ pkg/helm/helmcli.go
@@ -282,7 +282,7 @@ func getChartStatus(releaseName string, namespace string) (string, error) {
 
 // getReleaseState extracts the helmRelease state from an "ls -o json" command for a specific helmRelease/namespace
 func getReleaseState(releaseName string, namespace string) (string, error) {
-	releases, err := getReleases(namespace)
+	releases, err := GetReleases(namespace)
 	if err != nil {
 		if strings.Contains(err.Error(), "not found") {
 			return ChartNotFound, nil
@@ -343,7 +343,7 @@ func GetReleaseValues(log vzlog.VerrazzanoLogger, valueKeys []string, releaseNam
 
 // getReleaseAppVersion extracts the helmRelease app_version from a "ls -o json" command for a specific helmRelease/namespace
 func getReleaseAppVersion(releaseName string, namespace string) (string, error) {
-	releases, err := getReleases(namespace)
+	releases, err := GetReleases(namespace)
 	if err != nil {
 		if err.Error() == ChartNotFound {
 			return ChartNotFound, nil
@@ -362,7 +362,7 @@ func getReleaseAppVersion(releaseName string, namespace string) (string, error) 
 	return strings.TrimSpace(status), nil
 }
 
-func getReleases(namespace string) ([]*release.Release, error) {
+func GetReleases(namespace string) ([]*release.Release, error) {
 	settings := cli.New()
 	settings.SetNamespace(namespace)
 	actionConfig, err := actionConfigFn(vzlog.DefaultLogger(), settings, namespace)
